+++ src/constants.js
@@ -1,4 +1,3 @@
-export const MAX_CAR_BYTES_IN_MEMORY = 1024 * 1024 * 5
 export const CAR_CODE = 0x0202
 
 export const RATE_LIMIT_EXCEEDED = {